Set default audio volume

+++ Assets/_Assets/Scripts/Services/Audio/AudioSettingsData.cs
@@ -5,7 +5,7 @@
         public float VFXVolume;
         public float MusicVolume;
 
-        public AudioSettingsData(int vfxVolume, int musicVolume)
+        public AudioSettingsData(float vfxVolume, float musicVolume)
         {
             VFXVolume = vfxVolume;
             MusicVolume = musicVolume;
